pyproject.toml
[project]
name = "punjabi-lyrics"
version = "0.1.0"
description = "Extract and translate Punjabi song lyrics from MP3 files"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "google-generativeai>=0.8.0",
    "pydub>=0.25.1",
    "librosa>=0.10.0",
    "soundfile>=0.12.1",
    "indic-transliteration>=2.3.0",
    "googletrans==4.0.0-rc1",
    "python-dotenv>=1.0.0",
]
